feat(sandbox): improvements to recursive sandbox and frost
This changes the way sandbox and frost work together, ensuring they are both recursive and applied at all times, and ensuring that changes to Evolver data always happens via frost processes. It also ensures that changes are not applied until the completion of an evolution.

BREAKING CHANGE: Evolvers do not modify data until the end, all levels are always frosted