fix(sandbox): fix cement for nested sandboxes, enhance isSandboxProxy to account for inelligible properties
Cement was failing if the root object wasn't a sandbox, even if a nested object was. This was partly becuase the isSandboxProxy function was failing to appropriately detect nested Sandboxes if they were empty objects or otherwise were potentially elligible to BE sandboxes, but weren't detectable (e.g. an object with no keys). In that situation, although the object is elligible to be a sandbox, and contains no sandboxes, it still returned true because we didn't account for that situation.