fix(sandbox.defrost): allow options to determine whether defrosting a non-frosty object should error
User can specify `defrost(target, {mode: "strict"})`, which will throw if the target is not frosty (recursively). It does not necessarily cause problems to defrost a non-frosty object, at least not from sandbox's perspective, but it can be a good indicator of problems if you want to ensure your objects truly are frosty all the way down. If you hit the `defrost called on non-frost object` error, it's either due to a bug in sandbox, or due to manual manipulation which is incompatible with the limitations of sandbox.